pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adversarial-reasoning"
version = "0.1.0"
description = "White-box image-space adversarial attacks on medical imaging VLM agents"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Kosmas Apostolidis", email = "cosmasapostolidis@gmail.com" },
]
keywords = [
    "adversarial-attacks",
    "vision-language-models",
    "medical-imaging",
    "agentic-ai",
    "tool-use",
    "benchmark",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML
    "torch>=2.6.0",
    "transformers>=4.45",
    "accelerate>=1.1",
    "bitsandbytes>=0.43",
    "safetensors>=0.4",
    # Adversarial libraries
    # `torchattacks` dropped: its 3.5.x wheels pin `requests~=2.25.1` which is
    # irreconcilable with modern `datasets` / `smolagents`. Our PGD wrapper is
    # self-contained (src/adversarial_reasoning/attacks/pgd.py) and ART covers
    # APGD / AutoAttack.
    "adversarial-robustness-toolbox>=1.20",
    # Agent framework
    "smolagents>=1.4",
    # Inference / serving
    "ollama>=0.4",
    # Image IO
    "pillow>=10.0",
    "opencv-python>=4.9",
    # Medical image IO
    "pydicom>=3.0",
    "nibabel>=5.2",
    "SimpleITK>=2.3",
    # Data + stats
    "datasets>=3.0",
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.13",
    "scikit-learn>=1.5",
    "statsmodels>=0.14",
    # Utility
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "rich>=13.7",
    "tenacity>=9.0",
    "python-Levenshtein>=0.25",
    # Figures
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-xdist>=3.6",
    "pre-commit>=3.7",
    "ruff>=0.6",
    "mypy>=1.11",
    "ipykernel>=6.29",
]

[project.urls]
Homepage = "https://github.com/KosmasApostolidis/adversarial-reasoning-attacks"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
adversarial_reasoning = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
]
ignore = ["E501"]  # handled by formatter

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
files = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
markers = [
    "slow: marks tests as slow (GPU/model loading)",
    "gpu: marks tests that require a CUDA GPU",
    "smoke: marks tests as smoke tests (quick sanity)",
]
